refactor feature flags and vector search page
